types

+++ graph_gen/graph_generator.h
@@ -3,10 +3,10 @@
 #include <vector>
 #include <iostream>
 #include "../graphs.h"
-adjacencyList generate_graph_list(int nodes,int min_neighbors,int max_neighbors,int seed);
+adjacencyList generate_graph_list(unsigned long nodes,unsigned long min_neighbors,unsigned long max_neighbors,int seed);
 adjacencyList generate_graph_list_from_text();
 adjacencyList generate_graph_list_from_textfile(std::string textfile);
-adjancencyMatrix generate_graph_matrix(int nodes,int min_neighbors,int max_neighbors,int seed);
+adjancencyMatrix generate_graph_matrix(unsigned long nodes,unsigned long min_neighbors,unsigned long max_neighbors,int seed);
 void print_graph(adjacencyList g);
 void print_graph(adjancencyMatrix g);
